chore(license): make quota params longs instead of ints
ORB-421

+++ licensing/license-api/src/main/java/com/orbitalhq/licensing/License.kt
@@ -79,7 +79,7 @@ enum class LicensedEdition(val enabledFeatures: List<LicensedFeature>) {
 
 data class UsageQuota(
    val capability: MeteredCapability,
-   val limit: Int
+   val limit: Long
 )
 
 enum class MeteredCapability {
